# backend/docs/swagger.yaml
definitions:
...
  dto.ProductDTO:
    properties:
      manufacturer:
        type: string
      name:
        type: string
      product_id:
        type: string
      type_name:
        type: string
      volume:
        type: number
      weight:
        type: number
    type: object
  dto.ProductSaleDTO:
    properties:
      product_id:
        type: string
      product_name:
        type: string
      product_sales_id:
        type: string
      quantity:
        type: number
      sale_datetime:
        type: string
      sale_id:
        type: string
      sale_price:
        type: number
    type: object
...
  models.Product:
    properties:
      manufacturer:
        type: string
      name:
        type: string
      product_id:
        type: string
      type:
        $ref: '#/definitions/models.ProductType'
      type_id:
        type: string
      volume:
        type: number
      weight:
        type: number
    type: object
  models.ProductSale:
    properties:
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      product_sales_id:
        type: string
      quantity:
        type: number
      sale:
        $ref: '#/definitions/models.Sale'
      sale_id:
        type: string
      sale_price:
        type: number
    type: object
...
  models.ProductType:
    properties:
      for_adult:
        type: boolean
      name:
        type: string
      type_id:
        type: string
    type: object
  models.Sale:
    properties:
      datetime:
        type: string
      discount:
        type: number
      full_cost:
        type: number
      sale_id:
        type: string
      total_cost:
        type: number
    type: object
...
  response.Response:
    properties:
      error:
        description: Error if response is error
        type: string
      status:
        description: Response status
        type: string
    type: object
info:
  contact: {}
paths:
...
  /product-sales:
    get:
      description: Returns a list of products in a sale.
      parameters:
      - description: Sale ID
        format: uuid
        in: query
        name: sale_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductSaleDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get products in sale
      tags:
      - product-sales
    post:
      consumes:
      - application/json
      description: Adds a product to a sale.
      parameters:
      - description: Product sale
        in: body
        name: productSale
        required: true
        schema:
          $ref: '#/definitions/models.ProductSale'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductSale'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create product in sale
      tags:
      - product-sales
  /product-sales/{id}:
    delete:
      description: Removes a product from a sale by ID.
      parameters:
      - description: Product sale ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete product from sale
      tags:
      - product-sales
    put:
      consumes:
      - application/json
      description: Updates a product in a sale.
      parameters:
      - description: Product sale ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Product sale
        in: body
        name: productSale
        required: true
        schema:
          $ref: '#/definitions/models.ProductSale'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update product in sale
      tags:
      - product-sales
...
  /products:
    get:
      description: Returns a list of products filtered by type and search string.
      parameters:
      - description: Product type ID
        format: uuid
        in: query
        name: type_id
        type: string
      - description: Search string
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Creates a new product.
      parameters:
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Create product
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes a product by ID.
      parameters:
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Updates an existing product.
      parameters:
      - description: Product ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update product
      tags:
      - products
...
swagger: "2.0"